requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
kaggle==1.5.16
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
db-dtypes
docker
pytest
pandas
apache-airflow
timedelta